fix warnings/errors on 64-bit arch

+++ src/getopt.c
@@ -673,7 +673,7 @@ _getopt_internal (
       /* Test all long options for either exact match
          or abbreviated matches.  */
       for (p = longopts, option_index = 0; p->name; p++, option_index++)
-        if (!strncmp (p->name, nextchar, (unsigned)(nameend - nextchar)))
+        if (!strncmp (p->name, nextchar, (size_t)(nameend - nextchar)))
           {
             if ((unsigned int) (nameend - nextchar)
                 == (unsigned int) strlen (p->name))
@@ -1011,7 +1011,7 @@ _getopt_internal (
         /* Test all long options for either exact match
            or abbreviated matches.  */
         for (p = longopts, option_index = 0; p->name; p++, option_index++)
-          if (!strncmp (p->name, nextchar, (unsigned)(nameend - nextchar)))
+          if (!strncmp (p->name, nextchar, (size_t)(nameend - nextchar)))
             {
               if ((unsigned int) (nameend - nextchar) == strlen (p->name))
                 {
